requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
numpy>=1.26.0
